fix: Section 4,5 병합 및 재배치
