+++ src/pages/section/index.js
@@ -4,7 +4,6 @@ import Section3 from './Section3';
 import Section4 from './Section4';
 import Section5 from './Section5';
 import Section6 from './Section6';
-import Section7 from './Section7';
 
 export const Section = {
   Section1,
@@ -13,5 +12,4 @@ export const Section = {
   Section4,
   Section5,
   Section6,
-  Section7,
 };
